moved autologout to components folder
